+++ src/services/guilds/:guildId/delete.ts
